fixed some compiler warning

+++ lcadviewer/helpers/snapmanagerimpl.cpp
@@ -46,8 +46,8 @@ void SnapManagerImpl::on_mouseMoveEvent(const MouseMoveEvent& event) {
     if (_entities.size() > 1) {
         qSort(_entities.begin() , _entities.end(), lc::EntityDistance::sortAscending);
 
-        for (int a = 0; a < _entities.size(); a++) {
-            for (int b = a + 1; b < _entities.size(); b++) {
+        for (size_t a = 0; a < _entities.size(); a++) {
+            for (size_t b = a + 1; b < _entities.size(); b++) {
                 lc::CADEntity_CSPtr i1 = _entities.at(a).entity();
                 lc::CADEntity_CSPtr i2 = _entities.at(b).entity();
